// src/types/report.ts
export interface ReportType {
  id: number;
  kondisi: string;
  status_umur: string;
  kebun: string;
  kkl_kebun: string;
  afd: string;
  tahun_tanam: number;
  no_blok: string;
  luas: number;
  jlh_pokok: number;
  pkk_ha: number;
  r: number;
  warna: string;
  bulan : string;
  tahun : number;
  createdAt : string;
}